Move `parseComments` to a separated module
